Reduce enum noise
- The enum implementation that was generated shares a lot of boilerplate
among other enums and generally results in an implementation that has a
lot of noise around it compared to what we want to achieve. To reduce
the visual clutter in generated enums much of the functionality was
moved out into an EnumTrait.

- Instead of returning an arbitrary toString() we are now simply
returning the string representation (i.e. method name used to
instantiate your enum instance) of the enum that was generated. This is
more inline with our Java inspiration and is also more intuitive to use
in places where you might want to show or use the string representation.