Add DBAL Type, major internal refactor
- Implements functionality to generate a Doctrine DBAL Type based off of
a generated Enum. This allows the persistence of the string
representation of an Enum within Doctrine entities.

- Represents a major internal refactor that moves code that consuming
clients should not interact with into the Cspray\Yape\Internal
namespace. Many changes were made to the code within this namespace to
support the new DBAL Type and reuse as much code as was feasible.
Consuming clients SHOULD NOT use or expect strict BC between versions
for code within this namespace. Generally speaking the only code your
applications should use from this library are the implementations that
live under Cspray\Yape.