Improve adding multiple dependencies on Enum
Previously to properly set the enum value either required setting the
property directly, which we generally do not recommend or required extra
boilerplate in the Enum to rename the constructor to something else.
Instead of potentially having that "something else" be an arbitrary
potentially different name for each implementation we added an
explicit setEnumValue method that reduces amount of boilerplate
necessary to add multiple dependencies.